Inline @shared team schemas into route-local files
