--- src/routes/@shared/index.ts
@@ -1,21 +1,0 @@
-export { rules } from '../rules'
-
-export {
-  type CreateTeamBody,
-  createTeamBodySchema,
-  type CreateTeamCtx,
-  type GetTeamsCtx,
-  type GetTeamsQuery,
-  getTeamsQuerySchema,
-  type InviteMemberBody,
-  inviteMemberBodySchema,
-  type InviteMemberCtx,
-  type TeamParams as TeamInvitesParams,
-  teamParamsSchema as teamInvitesParamsSchema,
-  type TeamParams,
-  type TeamParamsCtx,
-  teamParamsSchema,
-  type UpdateTeamBody,
-  updateTeamBodySchema,
-  type UpdateTeamCtx,
-} from './team-schema'
